refactor: adjust backfill script
